Move to CherryPy 6.0.2, official distribution
