+++ cherrypy/LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2004-2015, CherryPy Team (team@cherrypy.org)
+Copyright (c) 2004-2016, CherryPy Team (team@cherrypy.org)
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
