+++ cherrypy/VERSION.txt
@@ -1,5 +1,4 @@
-CherryPy 5.1.0 Official distribution: https://pypi.python.org/packages/source/C/CherryPy/CherryPy-5.1.0.tar.gz
+CherryPy 6.0.2 Official distribution: https://pypi.python.org/packages/source/C/CherryPy/CherryPy-6.0.2.tar.gz
 The folders 'tutorial', 'test' and 'scaffold' have been removed.
 This file has been added.
-A patch is required to enable proper Bonjour/Zeroconfig support.
