+++ cherrypy/_cpcompat.py
@@ -7,7 +7,7 @@ preferring a newer idiom, sometimes an older one, and sometimes a custom one.
 In particular, Python 2 uses str and '' for byte strings, while Python 3
 uses str and '' for unicode strings. We will call each of these the 'native
 string' type for each version. Because of this major difference, this module
-provides new 'bytestr', 'unicodestr', and 'nativestr' attributes, as well as
+provides
 two functions: 'ntob', which translates native strings (of type 'str') into
 byte strings regardless of Python version, and 'ntou', which translates native
 strings to unicode strings. This also provides a 'BytesIO' name for dealing
@@ -20,11 +20,9 @@ import re
 import sys
 import threading
 
-if sys.version_info >= (3, 0):
-    py3k = True
-    bytestr = bytes
-    unicodestr = str
-    nativestr = unicodestr
+import six
+
+if six.PY3:
     basestring = (bytes, str)
 
     def ntob(n, encoding='ISO-8859-1'):
@@ -49,16 +47,8 @@ if sys.version_info >= (3, 0):
         if isinstance(n, bytes):
             return n.decode(encoding)
         return n
-    # type("")
-    from io import StringIO
-    # bytes:
-    from io import BytesIO as BytesIO
 else:
     # Python 2
-    py3k = False
-    bytestr = str
-    unicodestr = unicode
-    nativestr = bytestr
     basestring = basestring
 
     def ntob(n, encoding='ISO-8859-1'):
@@ -96,18 +86,10 @@ else:
         if isinstance(n, unicode):
             return n.encode(encoding)
         return n
-    try:
-        # type("")
-        from cStringIO import StringIO
-    except ImportError:
-        # type("")
-        from StringIO import StringIO
-    # bytes:
-    BytesIO = StringIO
 
 
 def assert_native(n):
-    if not isinstance(n, nativestr):
+    if not isinstance(n, str):
         raise TypeError("n must be a native str (got %s)" % type(n).__name__)
 
 try:
@@ -122,12 +104,12 @@ except ImportError:
 
 def base64_decode(n, encoding='ISO-8859-1'):
     """Return the native string base64-decoded (as a native string)."""
-    if isinstance(n, unicodestr):
+    if isinstance(n, six.text_type):
         b = n.encode(encoding)
     else:
         b = n
     b = _base64_decodebytes(b)
-    if nativestr is unicodestr:
+    if str is six.text_type:
         return b.decode(encoding)
     else:
         return b
@@ -221,7 +203,7 @@ except ImportError:
     from http.server import BaseHTTPRequestHandler
 
 # Some platforms don't expose HTTPSConnection, so handle it separately
-if py3k:
+if six.PY3:
     try:
         from http.client import HTTPSConnection
     except ImportError:
@@ -300,7 +282,7 @@ except ImportError:
         def _json_encode(s):
             raise ValueError('No JSON library is available')
 finally:
-    if json and py3k:
+    if json and six.PY3:
         # The two Python 3 implementations (simplejson/json)
         # outputs str. We need bytes.
         def json_encode(value):
