+++ cherrypy/_cpdispatch.py
@@ -422,7 +422,7 @@ class Dispatcher(object):
                     object_trail.insert(
                         i + 1, ["default", defhandler, conf, segleft])
                     request.config = set_conf()
-                    # See https://bitbucket.org/cherrypy/cherrypy/issue/613
+                    # See https://github.com/cherrypy/cherrypy/issues/613
                     request.is_index = path.endswith("/")
                     return defhandler, fullpath[fullpath_len - segleft:-1]
 
@@ -675,7 +675,7 @@ def VirtualHost(next_dispatcher=Dispatcher(), use_x_forwarded_host=True,
         result = next_dispatcher(path_info)
 
         # Touch up staticdir config. See
-        # https://bitbucket.org/cherrypy/cherrypy/issue/614.
+        # https://github.com/cherrypy/cherrypy/issues/614.
         section = request.config.get('tools.staticdir.section')
         if section:
             section = section[len(prefix):]
