+++ cherrypy/_cpserver.py
@@ -1,10 +1,10 @@
 """Manage HTTP servers with CherryPy."""
 
-import warnings
+import six
 
 import cherrypy
 from cherrypy.lib import attributes
-from cherrypy._cpcompat import basestring, py3k
+from cherrypy._cpcompat import basestring
 
 # We import * because we want to export check_port
 # et al as attributes of this module.
@@ -61,11 +61,11 @@ class Server(ServerAdapter):
 
     socket_timeout = 10
     """The timeout in seconds for accepted connections (default 10)."""
-    
+
     accepted_queue_size = -1
     """The maximum number of requests which will be queued up before
     the server refuses to accept it (default -1, meaning no limit)."""
-    
+
     accepted_queue_timeout = 10
     """The timeout in seconds for attempting to add a request to the
     queue when the queue is full (default 10)."""
@@ -113,7 +113,7 @@ class Server(ServerAdapter):
     ssl_private_key = None
     """The filename of the private key to use with SSL."""
 
-    if py3k:
+    if six.PY3:
         ssl_module = 'builtin'
         """The name of a registered SSL adaptation module to use with
         the builtin WSGI server. Builtin options are: 'builtin' (to
