+++ cherrypy/_cptools.py
@@ -26,6 +26,7 @@ import sys
 import warnings
 
 import cherrypy
+from cherrypy._helper import expose
 
 
 def _getargs(func):
@@ -113,10 +114,10 @@ class Tool(object):
 
         For example::
 
+            @expose
             @tools.proxy()
             def whats_my_base(self):
                 return cherrypy.request.base
-            whats_my_base.exposed = True
         """
         if args:
             raise TypeError("The %r Tool does not accept positional "
@@ -171,12 +172,12 @@ class HandlerTool(Tool):
                 nav = tools.staticdir.handler(section="/nav", dir="nav",
                                               root=absDir)
         """
+        @expose
         def handle_func(*a, **kw):
             handled = self.callable(*args, **self._merged_args(kwargs))
             if not handled:
                 raise cherrypy.NotFound()
             return cherrypy.serving.response.body
-        handle_func.exposed = True
         return handle_func
 
     def _wrapper(self, **kwargs):
@@ -271,7 +272,7 @@ class SessionTool(Tool):
         body. This is off by default for safety reasons; for example,
         a large upload would block the session, denying an AJAX
         progress meter
-        (`issue <https://bitbucket.org/cherrypy/cherrypy/issue/630>`_).
+        (`issue <https://github.com/cherrypy/cherrypy/issues/630>`_).
 
         When 'explicit' (or any other value), you need to call
         cherrypy.session.acquire_lock() yourself before using
@@ -365,6 +366,7 @@ class XMLRPCController(object):
     # would be if someone actually disabled the default_toolbox. Meh.
     _cp_config = {'tools.xmlrpc.on': True}
 
+    @expose
     def default(self, *vpath, **params):
         rpcparams, rpcmethod = _xmlrpc.process_body()
 
@@ -376,7 +378,7 @@ class XMLRPCController(object):
             body = subhandler(*(vpath + rpcparams), **params)
 
         else:
-            # https://bitbucket.org/cherrypy/cherrypy/issue/533
+            # https://github.com/cherrypy/cherrypy/issues/533
             # if a method is not found, an xmlrpclib.Fault should be returned
             # raising an exception here will do that; see
             # cherrypy.lib.xmlrpcutil.on_error
@@ -387,7 +389,6 @@ class XMLRPCController(object):
                         conf.get('encoding', 'utf-8'),
                         conf.get('allow_none', 0))
         return cherrypy.serving.response.body
-    default.exposed = True
 
 
 class SessionAuthTool(HandlerTool):
@@ -459,6 +460,13 @@ class Toolbox(object):
                 if settings.get("on", False):
                     tool = getattr(self, name)
                     tool._setup()
+
+    def register(self, point, **kwargs):
+        """Return a decorator which registers the function at the given hook point."""
+        def decorator(func):
+            setattr(self, kwargs.get('name', func.__name__), Tool(point, func, **kwargs))
+            return func
+        return decorator
 
 
 class DeprecatedTool(Tool):
