+++ cherrypy/_cptree.py
@@ -2,8 +2,10 @@
 
 import os
 
+import six
+
 import cherrypy
-from cherrypy._cpcompat import ntou, py3k
+from cherrypy._cpcompat import ntou
 from cherrypy import _cpconfig, _cplogging, _cprequest, _cpwsgi, tools
 from cherrypy.lib import httputil
 
@@ -261,7 +263,7 @@ class Tree(object):
         # to '' (some WSGI servers always set SCRIPT_NAME to '').
         # Try to look up the app using the full path.
         env1x = environ
-        if environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
+        if six.PY2 and environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
             env1x = _cpwsgi.downgrade_wsgi_ux_to_1x(environ)
         path = httputil.urljoin(env1x.get('SCRIPT_NAME', ''),
                                 env1x.get('PATH_INFO', ''))
@@ -274,26 +276,12 @@ class Tree(object):
 
         # Correct the SCRIPT_NAME and PATH_INFO environ entries.
         environ = environ.copy()
-        if not py3k:
-            if environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
-                # Python 2/WSGI u.0: all strings MUST be of type unicode
-                enc = environ[ntou('wsgi.url_encoding')]
-                environ[ntou('SCRIPT_NAME')] = sn.decode(enc)
-                environ[ntou('PATH_INFO')] = path[
-                    len(sn.rstrip("/")):].decode(enc)
-            else:
-                # Python 2/WSGI 1.x: all strings MUST be of type str
-                environ['SCRIPT_NAME'] = sn
-                environ['PATH_INFO'] = path[len(sn.rstrip("/")):]
+        if six.PY2 and environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
+            # Python 2/WSGI u.0: all strings MUST be of type unicode
+            enc = environ[ntou('wsgi.url_encoding')]
+            environ[ntou('SCRIPT_NAME')] = sn.decode(enc)
+            environ[ntou('PATH_INFO')] = path[len(sn.rstrip("/")):].decode(enc)
         else:
-            if environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
-                # Python 3/WSGI u.0: all strings MUST be full unicode
-                environ['SCRIPT_NAME'] = sn
-                environ['PATH_INFO'] = path[len(sn.rstrip("/")):]
-            else:
-                # Python 3/WSGI 1.x: all strings MUST be ISO-8859-1 str
-                environ['SCRIPT_NAME'] = sn.encode(
-                    'utf-8').decode('ISO-8859-1')
-                environ['PATH_INFO'] = path[
-                    len(sn.rstrip("/")):].encode('utf-8').decode('ISO-8859-1')
+            environ['SCRIPT_NAME'] = sn
+            environ['PATH_INFO'] = path[len(sn.rstrip("/")):]
         return app(environ, start_response)
