+++ cherrypy/_cpwsgi.py
@@ -8,9 +8,12 @@ still be translatable to bytes via the Latin-1 encoding!"
 """
 
 import sys as _sys
+import io
+
+import six
 
 import cherrypy as _cherrypy
-from cherrypy._cpcompat import BytesIO, bytestr, ntob, ntou, py3k, unicodestr
+from cherrypy._cpcompat import ntob, ntou
 from cherrypy import _cperror
 from cherrypy.lib import httputil
 from cherrypy.lib import is_closable_iterator
@@ -24,7 +27,7 @@ def downgrade_wsgi_ux_to_1x(environ):
     for k, v in list(environ.items()):
         if k in [ntou('PATH_INFO'), ntou('SCRIPT_NAME'), ntou('QUERY_STRING')]:
             v = v.encode(url_encoding)
-        elif isinstance(v, unicodestr):
+        elif isinstance(v, six.text_type):
             v = v.encode('ISO-8859-1')
         env1x[k.encode('ISO-8859-1')] = v
 
@@ -124,7 +127,7 @@ class InternalRedirector(object):
                 environ['REQUEST_METHOD'] = "GET"
                 environ['PATH_INFO'] = ir.path
                 environ['QUERY_STRING'] = ir.query_string
-                environ['wsgi.input'] = BytesIO()
+                environ['wsgi.input'] = io.BytesIO()
                 environ['CONTENT_LENGTH'] = "0"
                 environ['cherrypy.previous_request'] = ir.request
 
@@ -164,12 +167,12 @@ class _TrappedResponse(object):
         self.started_response = True
         return self
 
-    if py3k:
-        def __next__(self):
-            return self.trap(next, self.iter_response)
-    else:
-        def next(self):
-            return self.trap(self.iter_response.next)
+    def __next__(self):
+        return self.trap(next, self.iter_response)
+
+    # todo: https://pythonhosted.org/six/#six.Iterator
+    if six.PY2:
+        next = __next__
 
     def close(self):
         if hasattr(self.response, 'close'):
@@ -189,7 +192,7 @@ class _TrappedResponse(object):
             if not _cherrypy.request.show_tracebacks:
                 tb = ""
             s, h, b = _cperror.bare_error(tb)
-            if py3k:
+            if six.PY3:
                 # What fun.
                 s = s.decode('ISO-8859-1')
                 h = [(k.decode('ISO-8859-1'), v.decode('ISO-8859-1'))
@@ -227,7 +230,7 @@ class AppResponse(object):
     def __init__(self, environ, start_response, cpapp):
         self.cpapp = cpapp
         try:
-            if not py3k:
+            if six.PY2:
                 if environ.get(ntou('wsgi.version')) == (ntou('u'), 0):
                     environ = downgrade_wsgi_ux_to_1x(environ)
             self.environ = environ
@@ -236,22 +239,22 @@ class AppResponse(object):
             r = _cherrypy.serving.response
 
             outstatus = r.output_status
-            if not isinstance(outstatus, bytestr):
+            if not isinstance(outstatus, bytes):
                 raise TypeError("response.output_status is not a byte string.")
 
             outheaders = []
             for k, v in r.header_list:
-                if not isinstance(k, bytestr):
+                if not isinstance(k, bytes):
                     raise TypeError(
                         "response.header_list key %r is not a byte string." %
                         k)
-                if not isinstance(v, bytestr):
+                if not isinstance(v, bytes):
                     raise TypeError(
                         "response.header_list value %r is not a byte string." %
                         v)
                 outheaders.append((k, v))
 
-            if py3k:
+            if six.PY3:
                 # According to PEP 3333, when using Python 3, the response
                 # status and headers must be bytes masquerading as unicode;
                 # that is, they must be of type "str" but are restricted to
@@ -269,12 +272,12 @@ class AppResponse(object):
     def __iter__(self):
         return self
 
-    if py3k:
-        def __next__(self):
-            return next(self.iter_response)
-    else:
-        def next(self):
-            return self.iter_response.next()
+    def __next__(self):
+        return next(self.iter_response)
+
+    # todo: https://pythonhosted.org/six/#six.Iterator
+    if six.PY2:
+        next = __next__
 
     def close(self):
         """Close and de-reference the current request and response. (Core)"""
@@ -321,7 +324,7 @@ class AppResponse(object):
                                 self.environ.get('PATH_INFO', ''))
         qs = self.environ.get('QUERY_STRING', '')
 
-        if py3k:
+        if six.PY3:
             # This isn't perfect; if the given PATH_INFO is in the
             # wrong encoding, it may fail to match the appropriate config
             # section URI. But meh.
